# api.yaml
openapi: 3.0.1
info:
  title: weather_backend
  version: v1
paths:
...
  /weather:
    get:
      tags:
        - WeatherForecast
      responses:
        '429':
          description: Too Many Requests
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/WeatherData'
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherData'
            text/json:
              schema:
                $ref: '#/components/schemas/WeatherData'
  /weather/coordinates:
    get:
      tags:
        - WeatherForecast
      parameters:
        - name: latitude
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: longitude
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: units
          in: query
          schema:
            $ref: '#/components/schemas/WeatherUnit'
        - name: lang
          in: query
          schema:
            type: string
      responses:
        '429':
          description: Too Many Requests
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/WeatherData'
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherData'
            text/json:
              schema:
                $ref: '#/components/schemas/WeatherData'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '503':
          description: Service Unavailable
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /weather/city:
    get:
      tags:
        - WeatherForecast
      parameters:
        - name: cityName
          in: query
          required: true
          schema:
            type: string
        - name: stateCode
          in: query
          schema:
            type: string
        - name: countryCode
          in: query
          schema:
            type: string
        - name: units
          in: query
          schema:
            $ref: '#/components/schemas/WeatherUnit'
        - name: lang
          in: query
          schema:
            type: string
      responses:
        '429':
          description: Too Many Requests
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/WeatherData'
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherData'
            text/json:
              schema:
                $ref: '#/components/schemas/WeatherData'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '503':
          description: Service Unavailable
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/weather/city/{cityId}':
    get:
      tags:
        - WeatherForecast
      parameters:
        - name: cityId
          in: path
          required: true
          schema:
            type: number
            format: double
        - name: units
          in: query
          schema:
            $ref: '#/components/schemas/WeatherUnit'
        - name: lang
          in: query
          schema:
            type: string
      responses:
        '429':
          description: Too Many Requests
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/WeatherData'
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherData'
            text/json:
              schema:
                $ref: '#/components/schemas/WeatherData'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '503':
          description: Service Unavailable
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /weather/zip:
    get:
      tags:
        - WeatherForecast
      parameters:
        - name: zipCode
          in: query
          required: true
          schema:
            type: string
        - name: countryCode
          in: query
          schema:
            type: string
            default: us
        - name: units
          in: query
          schema:
            $ref: '#/components/schemas/WeatherUnit'
        - name: lang
          in: query
          schema:
            type: string
      responses:
        '429':
          description: Too Many Requests
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/WeatherData'
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherData'
            text/json:
              schema:
                $ref: '#/components/schemas/WeatherData'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '503':
          description: Service Unavailable
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
components:
  schemas:
...
    Cloud:
      type: object
      properties:
        all:
          type: integer
          format: int32
      additionalProperties: false
    Coordinate:
      type: object
      properties:
        lon:
          type: number
          format: double
        lat:
          type: number
          format: double
      additionalProperties: false
...
    MainWeather:
      type: object
      properties:
        temp:
          type: number
          format: double
        feels_like:
          type: number
          format: double
        temp_min:
          type: number
          format: double
        temp_max:
          type: number
          format: double
        pressure:
          type: integer
          format: int32
        humidity:
          type: integer
          format: int32
        sea_level:
          type: integer
          format: int32
          nullable: true
        grnd_level:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
    Rain:
      type: object
      properties:
        1h:
          type: number
          format: double
          nullable: true
        3h:
          type: number
          format: double
          nullable: true
      additionalProperties: false
...
    Snow:
      type: object
      properties:
        1h:
          type: number
          format: double
          nullable: true
        3h:
          type: number
          format: double
          nullable: true
      additionalProperties: false
...
    Weather:
      required:
        - description
        - icon
        - id
        - main
      type: object
      properties:
        id:
          type: integer
          format: int32
        main:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
      additionalProperties: false
    WeatherData:
      required:
        - clouds
        - coord
        - main
        - name
        - sys
        - weather
        - wind
      type: object
      properties:
        coord:
          $ref: '#/components/schemas/Coordinate'
        weather:
          type: array
          items:
            $ref: '#/components/schemas/Weather'
          nullable: true
        base:
          type: string
          nullable: true
        main:
          $ref: '#/components/schemas/MainWeather'
        visibility:
          type: integer
          format: int32
        wind:
          $ref: '#/components/schemas/Wind'
        clouds:
          $ref: '#/components/schemas/Cloud'
        rain:
          $ref: '#/components/schemas/Rain'
        snow:
          $ref: '#/components/schemas/Snow'
        dt:
          type: integer
          format: int32
        sys:
          $ref: '#/components/schemas/WeatherSystem'
        timezone:
          type: integer
          format: int32
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        cod:
          type: integer
          format: int32
      additionalProperties: false
    WeatherSystem:
      required:
        - country
        - sunrise
        - sunset
      type: object
      properties:
        type:
          type: integer
          format: int32
        id:
          type: integer
          format: int32
        country:
          type: string
          nullable: true
        sunrise:
          type: integer
          format: int32
        sunset:
          type: integer
          format: int32
      additionalProperties: false
    WeatherUnit:
      enum:
        - 0
        - 1
        - 2
      type: integer
      format: int32
    Wind:
      type: object
      properties:
        speed:
          type: number
          format: double
        deg:
          type: integer
          format: int32
        gust:
          type: number
          format: double
          nullable: true
      additionalProperties: false